@@ -19,9 +19,8 @@
         "resolve-url-loader": "^3.1.5",
         "sass": "^1.60.0",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.10.1",
-        "vue-template-compiler": "^2.6.12"
+        "vue": "^3.2.36",
+        "vue-loader": "^16.8.3"
     },
     "dependencies": {
         "jquery": "^3.6.4"
